pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "discernus"
version = "2.0.0"
description = "Discernus THIN v2.0 Research Platform"
readme = "README.md"
requires-python = ">=3.13"
license = {text = "GPL-3.0-or-later"}
authors = [
    {name = "Discernus Team"},
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: GNU General Public License v3 or later (GPLv3+)",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Text Processing :: Linguistic",
]

dependencies = [
    # Core dependencies
    "python-dotenv>=0.19.0",
    "gitpython>=3.1.0",
    "litellm>=1.7",
    "anthropic>=0.7.0",
    "requests>=2.28.0",
    "click>=8.0",
    "rich>=13.0",

    
    # Data processing and analysis
    "pandas>=2.2",
    "numpy>=1.26",
    "scipy>=1.13",
    
    # Jupyter and notebook support
    "jupyter>=1.0.0",
    "nbformat>=5.7.0",
    
    # Configuration and serialization
    "PyYAML>=6.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    
    # Text processing and NLP
    "nltk>=3.8",
    "textblob>=0.15.3",
    
    # Google Cloud Platform (Vertex AI)
    "google-cloud-aiplatform>=1.104.0",
    "google-auth>=2.40.0",
    
    # YouTube transcript extraction (for APDES corpus collection)
    "youtube-transcript-api>=0.6.2",
    "yt-dlp>=2025.6.9",
    "loguru>=0.7",
    "ratelimit>=2.2",
    "pingouin>=0.5",
    "statsmodels>=0.14",
    # Search and indexing (align with requirements.txt)
    "txtai>=5.0.0",
    "typesense>=0.20.0",
    "rank_bm25>=0.2.2",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-asyncio", 
    "black",
    "isort",
    "flake8",
]

academic = [
    # Additional scientific computing packages
    "statsmodels>=0.14.0",       # Statistical analysis
    "scikit-learn>=1.2.0",       # Machine learning
    "plotly>=5.14.0",            # Interactive visualizations
    "jupyter>=1.0.0",            # Notebook environment
]

[project.urls]
Homepage = "https://github.com/discernus/discernus"
Repository = "https://github.com/discernus/discernus.git"
Issues = "https://github.com/discernus/discernus/issues"

[project.scripts]
discernus = "discernus.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["discernus*"]

[tool.black]
line-length = 120
target-version = ['py38']

[tool.isort]
profile = "black"
line_length = 120

[tool.pytest.ini_options]
testpaths = ["discernus/tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"] 
python_functions = ["test_*"]
addopts = "-v" 